requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# NLP libraries
spacy>=3.3.0
nltk>=3.6.5
transformers>=4.18.0
sentence-transformers>=2.2.0
stanza>=1.3.0

# Graph processing
networkx>=2.6.0
neo4j>=4.4.3
pyvis>=0.2.0

# Machine learning
scikit-learn>=1.0.0
torch>=1.10.0

# Data ingestion
beautifulsoup4>=4.10.0
requests>=2.27.0
selenium>=4.1.0
pypdf2>=1.26.0
python-docx>=0.8.11
lxml>=4.8.0

# Visualization
matplotlib>=3.5.0
plotly>=5.6.0
streamlit>=1.8.0
streamlit-agraph>=0.0.35
streamlit-agraph>=0.0.35
pyvis>=0.2.0

# Utility
tqdm>=4.62.0
pyyaml>=6.0
pytest>=7.0.0